vision-app/requirements.txt
flask~=3.0.3
google-cloud-vision
opencv-python
google-cloud-storage